+++ nucliadb_models/nucliadb_models/__init__.py
@@ -17,6 +17,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
+import logging
 import re
 
 import pydantic
@@ -31,6 +32,8 @@ from .layout import *  # noqa
 from .link import *  # noqa
 from .metadata import *  # noqa
 from .text import *  # noqa
+
+logger = logging.getLogger("nucliadb_models")
 
 
 class SlugString(pydantic.ConstrainedStr):
