+++ nucliadb_protos/rust/src/nodereader.rs
@@ -287,7 +287,7 @@ pub mod node_reader_client {
     where
         T: tonic::client::GrpcService<tonic::body::BoxBody>,
         T::Error: Into<StdError>,
-        T::ResponseBody: Body<Data = Bytes> + Send + 'static,
+        T::ResponseBody: Default + Body<Data = Bytes> + Send + 'static,
         <T::ResponseBody as Body>::Error: Into<StdError> + Send,
     {
         pub fn new(inner: T) -> Self {
@@ -300,7 +300,6 @@ pub mod node_reader_client {
         ) -> NodeReaderClient<InterceptedService<T, F>>
         where
             F: tonic::service::Interceptor,
-            T::ResponseBody: Default,
             T: tonic::codegen::Service<
                 http::Request<tonic::body::BoxBody>,
                 Response = http::Response<
@@ -351,9 +350,9 @@ pub mod node_reader_client {
             &mut self,
             request: impl tonic::IntoRequest<super::super::noderesources::EmptyQuery>,
         ) -> Result<
-            tonic::Response<super::super::noderesources::ShardList>,
-            tonic::Status,
-        > {
+                tonic::Response<super::super::noderesources::ShardList>,
+                tonic::Status,
+            > {
             self.inner
                 .ready()
                 .await
@@ -557,9 +556,9 @@ pub mod node_reader_server {
             &self,
             request: tonic::Request<super::super::noderesources::EmptyQuery>,
         ) -> Result<
-            tonic::Response<super::super::noderesources::ShardList>,
-            tonic::Status,
-        >;
+                tonic::Response<super::super::noderesources::ShardList>,
+                tonic::Status,
+            >;
         async fn document_search(
             &self,
             request: tonic::Request<super::DocumentSearchRequest>,
