+++ nucliadb_protos/rust/src/resources.rs
@@ -506,6 +506,12 @@ pub struct FieldFile {
     pub language: ::prost::alloc::string::String,
     #[prost(string, tag="4")]
     pub password: ::prost::alloc::string::String,
+    #[prost(string, tag="5")]
+    pub url: ::prost::alloc::string::String,
+    #[prost(map="string, string", tag="6")]
+    pub headers: ::std::collections::HashMap<::prost::alloc::string::String, ::prost::alloc::string::String>,
+    #[prost(map="string, string", tag="7")]
+    pub cookies: ::std::collections::HashMap<::prost::alloc::string::String, ::prost::alloc::string::String>,
 }
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Entity {
