+++ nucliadb_writer/nucliadb_writer/api/models.py
@@ -22,7 +22,7 @@ from typing import Dict, Optional, Union
 from pydantic import BaseModel, validator
 
 import nucliadb_models as models
-from nucliadb_writer.processing import PushProcessingOptions
+from nucliadb_models.processing import PushProcessingOptions
 
 
 class CreateResourcePayload(BaseModel):
