+++ nucliadb_writer/nucliadb_writer/processing.py
@@ -20,72 +20,17 @@
 import datetime
 import uuid
 from contextlib import AsyncExitStack
-from enum import Enum
-from typing import TYPE_CHECKING, Any, Dict, List, Optional
+from typing import Any, Dict, List, Optional
 
 import aiohttp
 import jwt
 from nucliadb_protos.resources_pb2 import CloudFile
 from nucliadb_protos.resources_pb2 import FieldFile as FieldFilePB
-from pydantic import BaseModel, Field
 
 import nucliadb_models as models
 from nucliadb_utils.storages.storage import Storage
 from nucliadb_writer import logger
 from nucliadb_writer.exceptions import LimitsExceededError, SendToProcessError
-
-if TYPE_CHECKING:
-    SourceValue = CloudFile.Source.V
-else:
-    SourceValue = int
-
-
-class Source(SourceValue, Enum):  # type: ignore
-    HTTP = 0
-    INGEST = 1
-
-
-class PushProcessingOptions(BaseModel):
-    # Enable ML processing
-    ml_text: Optional[bool] = True
-
-
-class PushPayload(BaseModel):
-    # There are multiple options of payload
-    uuid: str
-    slug: Optional[str] = None
-    kbid: str
-    source: Optional[Source] = None
-    userid: str
-
-    genericfield: Dict[str, models.Text] = {}
-
-    # New File
-    filefield: Dict[str, str] = {}
-
-    # New Link
-    linkfield: Dict[str, models.LinkUpload] = {}
-
-    # Diff on Text Field
-    textfield: Dict[str, models.Text] = {}
-
-    # Diff on a Layout Field
-    layoutfield: Dict[str, models.LayoutDiff] = {}
-
-    # New conversations to process
-    conversationfield: Dict[str, models.PushConversation] = {}
-
-    # Only internal
-    partition: int
-
-    # List of available processing options (with default values)
-    processing_options: Optional[PushProcessingOptions] = Field(
-        default_factory=PushProcessingOptions
-    )
-
-
-class PushResponse(BaseModel):
-    seqid: Optional[int] = None
 
 
 class ProcessingEngine:
@@ -267,7 +212,9 @@ class ProcessingEngine:
 
         return jwt
 
-    async def send_to_process(self, item: PushPayload, partition: int) -> int:
+    async def send_to_process(
+        self, item: models.processing.PushPayload, partition: int
+    ) -> int:
         if self.dummy:
             self.calls.append(item.dict())
             return 1
