+++ nucliadb_writer/nucliadb_writer/resource/field.py
@@ -28,13 +28,13 @@ from nucliadb_ingest.fields.conversation import Conversation
 from nucliadb_ingest.orm.resource import Resource as ORMResource
 from nucliadb_models.common import FieldTypeName
 from nucliadb_models.conversation import PushConversation
+from nucliadb_models.processing import PushPayload
 from nucliadb_models.resource import FIELD_TYPES_MAP
 from nucliadb_protos import resources_pb2
 from nucliadb_utils.storages.storage import StorageField
 from nucliadb_utils.utilities import get_storage
 from nucliadb_writer.api.models import CreateResourcePayload, UpdateResourcePayload
 from nucliadb_writer.layouts import serialize_blocks
-from nucliadb_writer.processing import PushPayload
 from nucliadb_writer.utilities import get_processing
