+++ nucliadb_writer/nucliadb_writer/tests/test_resources.py
@@ -26,9 +26,9 @@ from httpx import AsyncClient
 
 import nucliadb_models
 from nucliadb_ingest.orm.resource import Resource
+from nucliadb_models.processing import PushPayload
 from nucliadb_models.resource import NucliaDBRoles
 from nucliadb_writer.api.v1.router import KB_PREFIX, RESOURCE_PREFIX, RESOURCES_PREFIX
-from nucliadb_writer.processing import PushPayload
 from nucliadb_writer.tests.test_fields import (
     TEST_CONVERSATION_PAYLOAD,
     TEST_DATETIMES_PAYLOAD,
